fix: refactored LTHButtonVariant to its own swift file
